+++ test/include/WinkTest/utils.h
@@ -5,7 +5,7 @@
 #include <gtest/gtest.h>
 
 #define ASSERT_ARRAY_EQ(length, expected, actual) \
-  for (int i = 0; i < length; i++)                \
+  for (size_t i = 0; i < length; i++)             \
     ASSERT_EQ(expected[i], actual[i]) << "Index: " << i;
 
 #endif  // TEST_INCLUDE_WINKTEST_UTILS_H_
